moving rewrite to rewrites module
